(pagination): Implement pagination destructure
- modify pagination component
- add pagination component to home page
- extract pagination view

[Starts #177933454]
